Refactor geometry selectors
